+++ src/portfolio/trade.rs
@@ -6,6 +6,9 @@ use crate::portfolio::Portfolio;
 use crate::traits::AsDataFrame;
 use crate::types::{Candle, ExecutedTrade, FailedTrade, FutureTrade, Side, Trade};
 use chrono::{NaiveDateTime, Utc};
+use rust_decimal::Decimal;
+use rust_decimal::prelude::{FromPrimitive, ToPrimitive};
+use rust_decimal_macros::dec;
 
 /// Interface methods for storing and retrieving trades, and determining when to trade
 pub trait TradeHandlers: PositionHandlers + AssetHandlers + CapitalHandlers {
@@ -18,7 +21,7 @@ pub trait TradeHandlers: PositionHandlers + AssetHandlers + CapitalHandlers {
     #[deprecated(note="Responsibility is moving to crate::PositionManager")]
     fn generate_buy_opt(&self, candle: &Candle) -> Option<FutureTrade>;
     #[deprecated(note="Responsibility is moving to crate::PositionManager")]
-    fn get_buy_cost(&self) -> f64;
+    fn get_buy_cost(&self) -> Decimal;
     fn get_last_trade(&self) -> Option<&ExecutedTrade>;
     #[deprecated(note="Responsibility is moving to crate::PositionManager")]
     fn able_to_buy(&self) -> bool;
@@ -65,10 +68,10 @@ impl TradeHandlers for Portfolio {
 
         if let Some(positions) = viable_positions {
             // the total quantity of assets to be sold
-            let quantity: f64 = positions.iter().map(|x| x.get_quantity()).sum();
+            let quantity = positions.iter().map(|x| x.get_quantity()).sum();
 
             // the total cost at which the assets were purchased
-            let cost: f64 = positions.iter().map(|x| x.get_cost()).sum();
+            let cost: Decimal = positions.iter().map(|x| x.get_cost()).sum();
 
             // calculate the value of the assets at the proposed rate
             let sell_value = match self.fee_calculator {
@@ -98,8 +101,8 @@ impl TradeHandlers for Portfolio {
     /// The amount of capital to use for a single buy trade
     ///
     /// This number is determined by the amount of capital available and the number of open positions.
-    fn get_buy_cost(&self) -> f64 {
-        self.available_capital() / self.available_open_positions() as f64
+    fn get_buy_cost(&self) -> Decimal {
+        self.available_capital() / Decimal::from(self.available_open_positions())
     }
 
     /// Get the most recent trade
@@ -151,12 +154,12 @@ impl TradeHandlers for Portfolio {
     }
 }
 
-fn calculate_buy_rate(candle: &Candle) -> f64 {
-    ((candle.close * 2.0) + candle.high + candle.open) / 4.0
+fn calculate_buy_rate(candle: &Candle) -> Decimal {
+    ((candle.close * dec!(2.0)) + candle.high + candle.open) / dec!(4.0)
 }
 
-fn calculate_sell_rate(candle: &Candle) -> f64 {
-    ((candle.close * 2.0) + candle.low + candle.open) / 4.0
+fn calculate_sell_rate(candle: &Candle) -> Decimal {
+    ((candle.close * dec!(2.0)) + candle.low + candle.open) / dec!(4.0)
 }
 
 #[cfg(test)]
@@ -164,21 +167,22 @@ mod tests {
     use crate::portfolio::{AssetHandlers, CapitalHandlers, Portfolio, TradeHandlers};
     use crate::types::{ExecutedTrade, FailedTrade, ReasonCode, Side, Trade};
     use chrono::{Duration, NaiveDateTime, Utc};
+    use rust_decimal_macros::dec;
 
     /// Test that a failed trade is correctly added to the portfolio storage.
     /// Since the failed storage is only for debugging and backtestesting, no other checks
     /// are necessary.
     #[test]
     fn test_add_failed_trade() {
-        let mut portfolio = Portfolio::new(100.0, 100.0, None);
+        let mut portfolio = Portfolio::new(dec!(100.0), dec!(100.0), None);
         assert!(portfolio.failed_trades.is_empty());
 
         // add a failed buy
         let trade = FailedTrade::new(
             ReasonCode::Unknown,
             Side::Buy,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
         );
         portfolio.add_failed_trade(trade);
@@ -188,8 +192,8 @@ mod tests {
         let trade = FailedTrade::new(
             ReasonCode::Unknown,
             Side::Sell,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
         );
         portfolio.add_failed_trade(trade);
@@ -200,55 +204,55 @@ mod tests {
     /// and that the capital and assets are updated appropriately
     #[test]
     fn test_add_executed_trade() {
-        let mut portfolio = Portfolio::new(200.0, 200.0, None);
+        let mut portfolio = Portfolio::new(dec!(200.0), dec!(200.0), None);
 
         // handle a buy
         let trade = ExecutedTrade::new_without_cost(
             "id".to_string(),
             Side::Buy,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
         );
         assert!(portfolio.executed_trades.is_empty());
-        assert_eq!(portfolio.available_capital(), 200.0);
-        assert_eq!(portfolio.get_assets(), 200.0);
+        assert_eq!(portfolio.available_capital(), dec!(200.0));
+        assert_eq!(portfolio.get_assets(), dec!(200.0));
 
         portfolio.add_executed_trade(trade);
         assert_eq!(portfolio.executed_trades.len(), 1);
         assert_eq!(portfolio.open_positions.len(), 1);
 
         // check that capital and assets are updated
-        assert_eq!(portfolio.available_capital(), 100.0);
-        assert_eq!(portfolio.get_assets(), 201.0);
+        assert_eq!(portfolio.available_capital(), dec!(100.0));
+        assert_eq!(portfolio.get_assets(), dec!(201.0));
 
         // handle a sell
         let trade = ExecutedTrade::new_without_cost(
             "id".to_string(),
             Side::Sell,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             (Utc::now() + Duration::seconds(1)).naive_utc()
         );
         portfolio.add_executed_trade(trade);
         assert_eq!(portfolio.executed_trades.len(), 2);
 
         // check that capital and assets are updated
-        assert_eq!(portfolio.available_capital(), 200.0);
-        assert_eq!(portfolio.get_assets(), 200.0);
+        assert_eq!(portfolio.available_capital(), dec!(200.0));
+        assert_eq!(portfolio.get_assets(), dec!(200.0));
         assert_eq!(portfolio.open_positions.len(), 0);
     }
 
     #[test]
     fn test_last_trade() {
-        let mut portfolio = Portfolio::new(200.0, 200.0, None);
+        let mut portfolio = Portfolio::new(dec!(200.0), dec!(200.0), None);
         assert!(portfolio.get_last_trade().is_none());
 
         let trade = ExecutedTrade::new_without_cost(
             "id".to_string(),
             Side::Buy,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             Utc::now().naive_utc(),
         );
         portfolio.add_executed_trade(trade);
@@ -259,8 +263,8 @@ mod tests {
         // append another trade and assert that the last trade is this new trade
         let id = "id".to_string();
         let side = Side::Sell;
-        let price = 121.0;
-        let quantity = 1.0;
+        let price = dec!(121.0);
+        let quantity = dec!(1.0);
         let time = Utc::now().naive_utc();
 
         let trade = ExecutedTrade::new_without_cost("id".to_string(), side, price, quantity, time);
@@ -280,16 +284,16 @@ mod tests {
     #[test]
     fn test_able_to_buy() {
         // test that we are able to buy when there are no open positions
-        let portfolio = Portfolio::new(100.0, 100.0, None);
+        let portfolio = Portfolio::new(dec!(100.0), dec!(100.0), None);
         assert!(portfolio.able_to_buy());
 
         // test that we are able to buy if the last trade is a sell
-        let mut portfolio = Portfolio::new(100.0, 100.0, None);
+        let mut portfolio = Portfolio::new(dec!(100.0), dec!(100.0), None);
         let trade = ExecutedTrade::new_without_cost(
             "id".to_string(),
             Side::Sell,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             Utc::now().naive_utc(),
         );
         portfolio.add_executed_trade(trade);
@@ -297,12 +301,12 @@ mod tests {
         assert!(portfolio.able_to_buy());
 
         // test that we are not able to buy if the last trade is a buy and the timeout has not expired
-        let mut portfolio = Portfolio::new(100.0, 100.0, None);
+        let mut portfolio = Portfolio::new(dec!(100.0), dec!(100.0), None);
         let trade = ExecutedTrade::new_without_cost(
             "id".to_string(),
             Side::Buy,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             Utc::now().naive_utc(),
         );
         portfolio.add_executed_trade(trade);
@@ -310,12 +314,12 @@ mod tests {
         assert!(!portfolio.able_to_buy());
 
         // test hat we are able to buy if the last trade is a buy and the timeout has expired
-        let mut portfolio = Portfolio::new(100.0, 100.0, None);
+        let mut portfolio = Portfolio::new(dec!(100.0), dec!(100.0), None);
         let trade = ExecutedTrade::new_without_cost(
             "id".to_string(),
             Side::Buy,
-            100.0,
-            1.0,
+            dec!(100.0),
+            dec!(1.0),
             Utc::now().naive_utc() - portfolio.timeout,
         );
         portfolio.add_executed_trade(trade);