+++ src/types/candles.rs
@@ -2,6 +2,8 @@ use crate::traits::AsDataFrame;
 use chrono::NaiveDateTime;
 use polars::frame::DataFrame;
 use polars::prelude::{NamedFrom, Series};
+use rust_decimal::Decimal;
+use rust_decimal::prelude::ToPrimitive;
 use serde::{Deserialize, Serialize};
 
 /// Abstracts a candlestick
@@ -10,11 +12,11 @@ pub struct Candle {
     #[serde(serialize_with = "crate::serialization::naive_dt_serializer")]
     #[serde(deserialize_with = "crate::serialization::naive_dt_deserializer")]
     pub time: NaiveDateTime,
-    pub open: f64,
-    pub high: f64,
-    pub low: f64,
-    pub close: f64,
-    pub volume: f64,
+    pub open: Decimal,
+    pub high: Decimal,
+    pub low: Decimal,
+    pub close: Decimal,
+    pub volume: Decimal,
 }
 
 impl<'de> Deserialize<'de> for Candle {
@@ -22,9 +24,9 @@ impl<'de> Deserialize<'de> for Candle {
     where
         D: serde::Deserializer<'de>,
     {
-        let arr = <[f64; 6]>::deserialize(deserializer)?;
+        let arr = <[Decimal; 6]>::deserialize(deserializer)?;
 
-        let time = NaiveDateTime::from_timestamp_opt(arr[0] as i64, 0).unwrap();
+        let time = NaiveDateTime::from_timestamp_opt(arr[0].to_i64().unwrap(), 0).unwrap();
         let open = arr[1];
         let high = arr[2];
         let low = arr[3];
@@ -46,11 +48,11 @@ impl AsDataFrame for Candle {
     fn as_dataframe(&self) -> DataFrame {
         DataFrame::new(vec![
             Series::new("time", vec![self.time]),
-            Series::new("open", vec![self.open]),
-            Series::new("high", vec![self.high]),
-            Series::new("low", vec![self.low]),
-            Series::new("close", vec![self.close]),
-            Series::new("volume", vec![self.volume]),
+            Series::new("open", vec![self.open.to_f64().unwrap()]),
+            Series::new("high", vec![self.high.to_f64().unwrap()]),
+            Series::new("low", vec![self.low.to_f64().unwrap()]),
+            Series::new("close", vec![self.close.to_f64().unwrap()]),
+            Series::new("volume", vec![self.volume.to_f64().unwrap()]),
         ])
         .unwrap()
     }
@@ -67,11 +69,11 @@ impl AsDataFrame for Vec<Candle> {
 
         for candle in self {
             time.push(candle.time);
-            open.push(candle.open);
-            high.push(candle.high);
-            low.push(candle.low);
-            close.push(candle.close);
-            volume.push(candle.volume);
+            open.push(candle.open.to_f64().unwrap());
+            high.push(candle.high.to_f64().unwrap());
+            low.push(candle.low.to_f64().unwrap());
+            close.push(candle.close.to_f64().unwrap());
+            volume.push(candle.volume.to_f64().unwrap());
         }
 
         DataFrame::new(vec![
@@ -91,17 +93,18 @@ mod test {
     use super::*;
     use chrono::Utc;
     use polars::prelude::AnyValue;
+    use rust_decimal_macros::dec;
 
     #[test]
     fn test_as_dataframe() {
         let time = NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap();
         let candle = Candle {
             time,
-            open: 1.0,
-            high: 2.0,
-            low: 3.0,
-            close: 4.0,
-            volume: 5.0,
+            open: dec!(1.0),
+            high: dec!(2.0),
+            low: dec!(3.0),
+            close: dec!(4.0),
+            volume: dec!(5.0),
         };
         let df = candle.as_dataframe();
         assert_eq!(df.shape(), (1, 6));
@@ -120,23 +123,23 @@ mod test {
         );
         assert_eq!(
             df.column("open").unwrap().get(0).unwrap(),
-            AnyValue::Float64(candle.open)
+            AnyValue::Float64(candle.open.to_f64().unwrap())
         );
         assert_eq!(
             df.column("high").unwrap().get(0).unwrap(),
-            AnyValue::Float64(candle.high)
+            AnyValue::Float64(candle.high.to_f64().unwrap())
         );
         assert_eq!(
             df.column("low").unwrap().get(0).unwrap(),
-            AnyValue::Float64(candle.low)
+            AnyValue::Float64(candle.low.to_f64().unwrap())
         );
         assert_eq!(
             df.column("close").unwrap().get(0).unwrap(),
-            AnyValue::Float64(candle.close)
+            AnyValue::Float64(candle.close.to_f64().unwrap())
         );
         assert_eq!(
             df.column("volume").unwrap().get(0).unwrap(),
-            AnyValue::Float64(candle.volume)
+            AnyValue::Float64(candle.volume.to_f64().unwrap())
         );
     }
 
@@ -145,14 +148,22 @@ mod test {
         // generate some candles
         let mut candles = Vec::new();
         for i in 0..10 {
+            let _i = Decimal::from(i);
             let time = NaiveDateTime::from_timestamp_opt(Utc::now().timestamp() + i, 0).unwrap();
+
+            let open = dec!(1.0);
+            let high = dec!(2.0);
+            let low = dec!(3.0);
+            let close = dec!(4.0);
+            let volume = dec!(5.0);
+
             let candle = Candle {
                 time,
-                open: 1.0 + i as f64,
-                high: 2.0 + i as f64,
-                low: 3.0 + i as f64,
-                close: 4.0 + i as f64,
-                volume: 5.0 + i as f64,
+                open: open + _i,
+                high: high + _i,
+                low: low + _i,
+                close: close + _i,
+                volume: volume + _i,
             };
             candles.push(candle);
         }
@@ -174,23 +185,23 @@ mod test {
             );
             assert_eq!(
                 df.column("open").unwrap().get(i).unwrap(),
-                AnyValue::Float64(candles[i].open)
+                AnyValue::Float64(candles[i].open.to_f64().unwrap())
             );
             assert_eq!(
                 df.column("high").unwrap().get(i).unwrap(),
-                AnyValue::Float64(candles[i].high)
+                AnyValue::Float64(candles[i].high.to_f64().unwrap())
             );
             assert_eq!(
                 df.column("low").unwrap().get(i).unwrap(),
-                AnyValue::Float64(candles[i].low)
+                AnyValue::Float64(candles[i].low.to_f64().unwrap())
             );
             assert_eq!(
                 df.column("close").unwrap().get(i).unwrap(),
-                AnyValue::Float64(candles[i].close)
+                AnyValue::Float64(candles[i].close.to_f64().unwrap())
             );
             assert_eq!(
                 df.column("volume").unwrap().get(i).unwrap(),
-                AnyValue::Float64(candles[i].volume)
+                AnyValue::Float64(candles[i].volume.to_f64().unwrap())
             );
         }
     }